% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnCommunityMeasures.R
\name{calc_Zi_Pi}
\alias{calc_Zi_Pi}
\title{Return in/out of community strength and Pi/Zi ()}
\usage{
calc_Zi_Pi(x.net, memberships)
}
\arguments{
\item{x.net}{a network object, in matrix format}

\item{memberships}{a vector of community memberships (e.g from igraph communties $memberhip)}
}
\value{
Pi -
Zi - a normalised measure of an individuals interactions within its subgroup
}
\description{
Return in/out of community strength and Pi/Zi ()
}
