% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetworks.R
\name{convert_AI}
\alias{convert_AI}
\title{Create association indexes}
\usage{
convert_AI(
  contacts_matrix = contacts_matrix,
  sampling_period = sampling_period,
  rownames = TRUE,
  diag = FALSE
)
}
\arguments{
\item{contacts_matrix}{dataframe of contacts in  the form ake the form of a  dataframe titled
id1, id2 and seconds, where the ids are  the individuals}

\item{sampling_period}{sampling period of the contact sensors. 20s for sheep work}

\item{rownames}{logical indicating if there are row/column names on the matrix}

\item{diag}{logical indicating if  the diagonal should  be included (i.e. are self loops possible). Default is false - self loops are not possible.}
}
\value{
new.mat - a new matrix consisting of the association indexes between individuals.
}
\description{
Create association indexes
}
\examples{
#'  #sort later
}
