% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetworks.R
\name{create_network_adjacency}
\alias{create_network_adjacency}
\title{Create_network_adjacency}
\usage{
create_network_adjacency(
  contacts,
  diag = FALSE,
  id1 = id1,
  id2 = id2,
  seconds = seconds
)
}
\arguments{
\item{contacts}{a dataframe of contacts in  the form ake the form of a  dataframe titled
id1, id2 and seconds, where the ids are  the individuals}

\item{diag}{logical indicating if  the diagonal should  be included. If false, will be NA, indicating no self loops}

\item{id1}{first id column}

\item{id2}{second id column}

\item{seconds}{length of contact (seconds)}
}
\value{
an adjacnecy matrix
}
\description{
Create_network_adjacency
}
\examples{
#' # add later
}
