% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetworks.R
\name{create_network_igraph}
\alias{create_network_igraph}
\title{Create network for igraph
This function takes a contacts list, and returns an igraph network object for  a  weighted network}
\usage{
create_network_igraph(contacts, id1 = id1, id2 = id2, seconds = seconds)
}
\arguments{
\item{contacts}{the contacts list should take the form of a  data frame titled
id1, id2 and seconds, where the ids are  the individuals}

\item{id1}{first id column}

\item{id2}{second id column}

\item{seconds}{length of contact (seconds)}
}
\value{
returns an igraph object
}
\description{
Create network for igraph
This function takes a contacts list, and returns an igraph network object for  a  weighted network
}
\examples{
# add later
}
