% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnCommunityMeasures.R
\name{return.community.measures}
\alias{return.community.measures}
\title{Title Network measures related to the communities}
\usage{
return.community.measures(x.net, community.net)
}
\arguments{
\item{x.net}{an network object, in matrix format}

\item{community.net}{an igraph communities object}
}
\value{
A dataframe of the measures that are related to the community formation.
Sub-group cohesion -  proportion of contacts that occur within sub-groups
mean_sub_group_cohesion = mean sub-group cohesion
var_sub_group_cohesion = variation in sub-group cohesion
In  and out of community strength for all individuals
mean_w_degree = mean weighted degree (assuming a weighted network)
var_w_degree = variance weighted degree (assuming a weighted network)
}
\description{
Title Network measures related to the communities
}
\examples{
example
}
