% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnCommunityMeasures.R
\name{return.community}
\alias{return.community}
\title{Title}
\usage{
return.community(network)
}
\arguments{
\item{network}{an igraph network object}
}
\value{
an igraph communities object
}
\description{
Title
}
\examples{
## place holder
}
