% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnNetworkMeasures.R
\name{return.lcs}
\alias{return.lcs}
\title{Return local cosine similarities between network at time 1 and time 2}
\usage{
return.lcs(m1, m2)
}
\arguments{
\item{m1}{network at time point 1}

\item{m2}{network at time point 2}
}
\value{
Vector of local csine similarities  for network 1 and 2
}
\description{
Return local cosine similarities between network at time 1 and time 2
}
\examples{
## fill in later
}
